+++ src/frontend/web_application/src/modules/draftMessage/components/DraftMessage/presenter.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import classnames from 'classnames';
 import { i18nMark, withI18n, Trans } from '@lingui/react';
 import {
-  Button, Icon, InputText, TextareaFieldGroup, TextFieldGroup, Link, Confirm, PlaceholderBlock,
+  Button, Icon, TextareaFieldGroup, TextFieldGroup, Link, Confirm, PlaceholderBlock,
   Callout, FieldErrors, Spinner,
 } from '../../../../components';
 import { LockedMessage } from '../../../../modules/encryption';
@@ -17,12 +17,13 @@ import RecipientList from '../RecipientList';
 import AttachmentManager from '../AttachmentManager';
 import IdentitySelector from '../IdentitySelector';
 import RecipientSelector from '../RecipientSelector';
+import QuickDraftForm from '../QuickDraftForm';
+import ToggleAdvancedFormButton from '../ToggleAdvancedFormButton';
 import { getIdentityProtocol } from '../../services/getIdentityProtocol';
 
 import './draft-message-quick.scss';
 import './draft-message-advanced.scss';
 import './draft-message-placeholder.scss';
-import './toggle-advanced-draft-button.scss';
 
 const PROTOCOL_EMAIL = 'email';
 
@@ -160,49 +161,6 @@ class DraftMessage extends Component {
       // eslint-disable-next-line react/no-did-update-set-state
       this.setState(prevState => this.constructor.genererateStateFromProps(this.props, prevState));
     }
-  }
-
-  getRecipientList = () => {
-    const { draftMessage } = this.props;
-
-    // participants may not be present when the draft is new, it is the responsibility of the
-    // backend to calculate what will be the participants for a reply
-    if (!draftMessage || !draftMessage.participants) {
-      return null;
-    }
-
-    const recipients = getRecipients(draftMessage);
-
-    if (!recipients) {
-      return '';
-    }
-
-    if (recipients.length > 3) {
-      return recipients.map(recipient => recipient.address).join(', ').concat('...');
-    }
-
-    return recipients.map(recipient => recipient.address).join(', ');
-  }
-
-  getQuickInputPlaceholder = () => {
-    const { i18n, draftMessage, availableIdentities } = this.props;
-
-    const [identityId] = draftMessage && draftMessage.user_identities ?
-      draftMessage.user_identities : [];
-    const { identifier } = availableIdentities
-      .find(identity => identity.identity_id === identityId) || {};
-
-    const recipientsList = this.getRecipientList();
-
-    if (draftMessage && draftMessage.discussion_id && recipientsList && identifier) {
-      return i18n._('draft-message.form.placeholder.quick-reply', { recipients: recipientsList, protocol: identifier }, { defaults: 'Quick reply to {recipients} from {protocol}' });
-    }
-
-    if (draftMessage && draftMessage.discussion_id && identifier) {
-      return i18n._('draft-message.form.placeholder.quick-reply-no-recipients', { identifier }, { defaults: 'Quick reply from {identifier}' });
-    }
-
-    return i18n._('draft-message.form.placeholder.quick-start', null, { defaults: 'Start a new discussion' });
   }
 
   getIdentity = ({ identityId }) => this.props.availableIdentities
@@ -472,24 +430,6 @@ class DraftMessage extends Component {
     );
   }
 
-  renderToggleAdvancedButton() {
-    const { i18n } = this.props;
-    const caretType = this.state.advancedForm ? 'caret-up' : 'caret-down';
-
-    return (
-      <Button
-        display="expanded"
-        shape="plain"
-        className="m-toggle-advanced-draft-button"
-        title={i18n._('draft-message.action.toggle-advanced', null, { defaults: 'Toggle advanced or quick message form' })}
-        onClick={this.handleToggleAdvancedForm}
-      >
-        <Icon type="envelope" />
-        <Icon type={caretType} />
-      </Button>
-    );
-  }
-
   renderRecipientList({ className } = {}) {
     const { canEditRecipients, isReply } = this.props;
 
@@ -531,8 +471,8 @@ class DraftMessage extends Component {
 
   renderQuick() {
     const {
-      className, i18n, draftFormRef, onFocus, scrollTarget: { forwardRef },
-      draftEncryption, isEncrypted, encryptionStatus,
+      className, draftFormRef, onFocus, scrollTarget: { forwardRef },
+      draftEncryption, isEncrypted, encryptionStatus, availableIdentities, draftMessage,
     } = this.props;
     const ref = (el) => {
       draftFormRef(el);
@@ -544,58 +484,26 @@ class DraftMessage extends Component {
     const canSend = this.getCanSend() && this.state.draftMessage.body.length > 0;
 
     return (
-      <div className={classnames(className, 'm-draft-message-quick')} ref={ref}>
-        <form onSubmit={this.handleSend}>
-          <div className={classnames(className, 'm-draft-message-quick__container')}>
-            <div className="m-draft-message-quick__toggle-advanced">
-              {this.renderToggleAdvancedButton()}
-            </div>
-            {
-              this.state.isLocked ?
-                <LockedMessage encryptionStatus={draftEncryption} />
-                : (
-                  <InputText
-                    className={classnames(
-                      'm-draft-message-quick__input',
-                      { 'm-draft-message-quick__input--encrypted': encryptionEnabled }
-                    )}
-                    onChange={this.handleChange}
-                    onFocus={onFocus}
-                    name="body"
-                    value={this.state.draftMessage.body}
-                    placeholder={this.getQuickInputPlaceholder()}
-                  />
-                )}
-            <div className={classnames(
-              'm-draft-message-quick__send',
-              {
-                'm-draft-message-quick__send--encrypted': encryptionEnabled,
-                'm-draft-message-quick__send--unencrypted': !encryptionEnabled,
-              }
-            )}
-            >
-              <Button
-                type="submit"
-                display="expanded"
-                shape="plain"
-                icon={this.state.isSending ? (<Spinner loading display="block" />) : 'paper-plane'}
-                title={i18n._('draft-message.action.send', null, { defaults: 'Send' })}
-                className={classnames(
-                  'm-draft-message-quick__send-button',
-                  {
-                    'm-draft-message-quick__send-button--encrypted': encryptionEnabled,
-                    'm-draft-message-quick__send-button--unencrypted': !encryptionEnabled,
-                  }
-                )}
-                disabled={!canSend}
-              />
-            </div>
-          </div>
-          <div className="m-draft-message-quick__encryption">
-            <Trans id={this.getEncryptionTranslation()} />
-          </div>
-        </form>
-      </div>
+      <QuickDraftForm
+        ref={ref}
+        className={className}
+        handleSend={this.handleSend}
+        handleChange={this.handleChange}
+        onFocus={onFocus}
+        draftEncryption={draftEncryption}
+        encryptionEnabled={encryptionEnabled}
+        isLocked={this.state.isLocked}
+        body={this.state.draftMessage.body}
+        canSend={canSend}
+        encryptionChildren={(
+          <Trans id={this.getEncryptionTranslation()} />
+        )}
+        availableIdentities={availableIdentities}
+        draftMessage={draftMessage}
+        isSending={this.state.isSending}
+        advancedForm={this.state.advancedForm}
+        handleToggleAdvancedForm={this.handleToggleAdvancedForm}
+      />
     );
   }
 
@@ -635,7 +543,12 @@ class DraftMessage extends Component {
     return (
       <div className={classnames(className, 'm-draft-message-advanced')} ref={ref}>
         <div className="m-draft-message-advanced__toggle-simple">
-          {isReply && errors.length === 0 && this.renderToggleAdvancedButton()}
+          {isReply && errors.length === 0 && (
+            <ToggleAdvancedFormButton
+              advancedForm={this.state.advancedForm}
+              handleToggleAdvancedForm={this.handleToggleAdvancedForm}
+            />
+          )}
         </div>
         <div className="m-draft-message-advanced__container">
           <IdentitySelector